Fixed DAOs

+++ server/api/src/main/java/de/sky/regular/income/api/Transaction.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 @Data
 public class Transaction {
     public UUID id;
-    public Double amount;
+    public Integer amount;
     public LocalDate date;
     public Boolean isPeriodic;
     public String reason;
